requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   langchain-community
    #   litellm
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
build==1.4.0
    # via pip-tools
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   ddgs
    #   duckduckgo-search
    #   litellm
    #   pip-tools
dataclasses-json==0.6.7
    # via langchain-community
ddgs==9.11.4
    # via -r requirements.in
distro==1.9.0
    # via openai
duckduckgo-search==8.1.1
    # via -r requirements.in
fastuuid==0.14.0
    # via litellm
filelock==3.25.2
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flashrank==0.2.10
    # via -r requirements.in
flatbuffers==25.12.19
    # via onnxruntime
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.2.0
    # via
    #   huggingface-hub
    #   torch
grpcio==1.78.0
    # via qdrant-client
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hf-xet==1.4.2
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   litellm
    #   openai
    #   qdrant-client
httpx-sse==0.4.3
    # via langchain-community
huggingface-hub==0.36.2
    # via
    #   sentence-transformers
    #   smolagents
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.1
    # via litellm
jinja2==3.1.6
    # via
    #   litellm
    #   smolagents
    #   torch
jiter==0.13.0
    # via openai
joblib==1.5.3
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.26.0
    # via litellm
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain-classic==1.0.3
    # via langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.2.19
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==1.1.1
    # via langchain-classic
langsmith==0.7.20
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
litellm==1.82.3
    # via -r requirements.in
lxml==6.0.2
    # via
    #   ddgs
    #   duckduckgo-search
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
marshmallow==3.26.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mdutils==1.8.1
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
networkx==3.6.1
    # via torch
numpy==2.4.2
    # via
    #   -r requirements.in
    #   flashrank
    #   langchain-community
    #   onnxruntime
    #   pandas
    #   qdrant-client
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
onnxruntime==1.24.3
    # via flashrank
openai==2.28.0
    # via litellm
orjson==3.11.7
    # via langsmith
packaging==26.0
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   onnxruntime
    #   transformers
    #   wheel
pandas==3.0.1
    # via -r requirements.in
pillow==12.1.1
    # via smolagents
pip-tools==7.5.3
    # via -r requirements.in
portalocker==3.2.0
    # via qdrant-client
primp==1.1.3
    # via
    #   ddgs
    #   duckduckgo-search
prompt-toolkit==3.0.52
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==7.34.0
    # via
    #   onnxruntime
    #   qdrant-client
pydantic==2.12.5
    # via
    #   langchain-classic
    #   langchain-core
    #   langsmith
    #   litellm
    #   openai
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via langchain-community
pyfiglet==1.0.4
    # via -r requirements.in
pygments==2.19.2
    # via rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   litellm
    #   pydantic-settings
    #   smolagents
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   langchain-classic
    #   langchain-community
    #   langchain-core
    #   transformers
qdrant-client==1.17.1
    # via -r requirements.in
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.2.28
    # via
    #   tiktoken
    #   transformers
requests==2.32.5
    # via
    #   -r requirements.in
    #   flashrank
    #   huggingface-hub
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   smolagents
    #   tiktoken
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
rich==14.3.3
    # via
    #   -r requirements.in
    #   smolagents
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.17.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.3.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smolagents==1.24.0
    # via -r requirements.in
sniffio==1.3.1
    # via openai
sqlalchemy==2.0.48
    # via
    #   langchain-classic
    #   langchain-community
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
tantivy==0.25.1
    # via -r requirements.in
tenacity==9.1.4
    # via
    #   langchain-community
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.12.0
    # via litellm
tokenizers==0.22.2
    # via
    #   flashrank
    #   litellm
    #   transformers
torch==2.10.0
    # via sentence-transformers
tqdm==4.67.3
    # via
    #   flashrank
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.57.6
    # via sentence-transformers
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   grpcio
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sentence-transformers
    #   sqlalchemy
    #   torch
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via
    #   qdrant-client
    #   requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
wcwidth==0.6.0
    # via prompt-toolkit
wheel==0.46.3
    # via pip-tools
xxhash==3.6.0
    # via langsmith
yarl==1.23.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
zstandard==0.25.0
    # via langsmith

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
